[FIX] sale_subscription/saas~15.3.1.1: set_recurring_id_in_sale_order
The issue occurs when a customer upgrades from version 15.0 to 18.0 and tries to
open an invoice from a sales order.
A pop-up message appear stating that the recurring_plan (plan_id) field is null.
This error occurs because the `has_recurring_line` field,
A has_recurring_line, was introduced in version 17.0.
This is a computed boolean field that is determined based on the product.
If a product has recurring_invoice set to True
then has_recurring_line is also True

In version 18.0, this field is linked to the plan_id field as a required value.
This means that if a sales order includes a product of type "recurring,"
the sales order must have a corresponding recurring_plan value.

closes odoo/upgrade#7286

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>