[IMP] l10n_br_test_avatax_sale/saas~18.3: Remove unused tests module
The l10n_br_test_avatax_sale module was kept after the refactor of
external_tax modules because the code of the module was no longer
necessary but the test cases still had use.

Later down the line, the l10n_br_avatax_sale module was created to
account for CNAE, operation types, and other new features.

As this module expands it makes sense to move the test code back to
it instead of keeping a separate test module.

See odoo/enterprise#81525

task-4569548

closes odoo/upgrade#7390

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>