[FIX] hr_recruitment/saas~18.2: move action to `hr_recruitment_sms`
Odoo: https://github.com/odoo/odoo/pull/194647

The module `sms` is getting installed when we install `hr_recruitment`
because of **auto_install=TRUE**. But we can uninstall `sms`
independently.

There were 2 action windows in `hr_recruitment` which were depending on
module `sms` but it gets problematic when client uninstalls `sms`.
1. `hr_recruitment.action_hr_applicant_mass_sms`[^1] was introduced
   in `saas~16.4`
2. `hr_recruitment.action_hr_candidate_mass_sms`[^2] was introduced
   in `18.0`.

For solving the dependency issue between `hr_recruitment` and `sms`, bridge
module `hr_recruitment_sms` was introduced in `saas~18.2`[^3] and the
action windows were moved to this module. But renaming
xmlids from `hr_recruitment` --> `hr_recruitment_sms` is missing.

So if we do this steps:
1. Install `hr_recruitment` in `18.0`
2. Uninstall `sms`
3. Upgrade to `saas~18.2`

Those action windows will be removed during the upgrade and recreated
which might be problem if they were being used by custom menu or view.

The reason it was made in `saas~18.2` is that modifying older stable
versions could potentially cause other issues.

In this commit, I renamed the xmlid of `action_hr_applicant_mass_sms`
if `hr_recruitment_sms` is installed and deleted it if not. The fix
is in `hr_recruitment` since the bridge module is new and can't run
scripts before loading. Additionally, the action window
`hr_recruitment.action_hr_candidate_mass_sms`[^4] is being removed anyway,
so I only renamed the first one.

[^1]: https://github.com/odoo/odoo/commit/5da9a87a8255fea1c05039304ed875f9135372f0
[^2]: https://github.com/odoo/odoo/blob/e5b2a1744ff2d2ab4fec8b79256519107e561bcd/addons/hr_recruitment/views/hr_candidate_views.xml#L220-L232
[^3]: https://github.com/odoo/odoo/pull/194647
[^4]: https://github.com/odoo/odoo/pull/189274

closes odoo/upgrade#7302

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>