[IMP] spreadsheet_edition/saas~18.3: drop missing sorted column
Master companion for a fix in 18.0

Steps: Go to a pivot view with:

- either a predefined sort on the action/filter.
- sort by a measure then uncheck the measure.

The `sortedColumn` field is still in the pivot definition but
the field is no longer in the `measures`.
It cannot work because the sorting is done client-side.
If the field is not part of the measures, we don't have the data
to sort...

In 18.0, the sorted column is ignored. Here in master, we drop it
from the core definition.

This upgrade PR cleans existing faulty revisions.

odoo/o-spreadsheet#5920 (clean fix for master)
odoo/odoo#193739 (stable compatible fix in 18.0)

closes odoo/upgrade#7305

Task: 4467262
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>