[FIX] documents/saas~17.3: add default value for new field color
In 17.3, new column color in documents facet model for tag category
but there is no default value set so for all color it's set
Null value.
when database upgraded from 17.0 to 18.0, their document tags
don't appear on their documents. The tags are still there, you
just can't see them.because your existing
tags do not have a color selected. Odoo treats tags with no
color as invisible in Kanban view (the "normal" view you were
referring to).so tag are not visible in kanban because
there are null value for color.
so need to set default value for color.

see:
https://github.com/odoo/enterprise/commit/721452b0e7497c31aa345d2920c93ce72ece2176

upg-2459016
unicorn-4604185
opw-4604391

closes odoo/upgrade#7337

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>