[FIX] account/saas~17.5.1.2: avoid deadlock detected
This is a follow-up to 3ece39f8406db368e8991e86be9b9e481ac336a2,
which did not solve the issue, because it wasn't fully understood, which
changed in the meantime:

ALTER INDEX and ALTER TABLE both acquire a variant of exclusive lock. This lock
is held until the end of the transaction. If in between the two, some
concurrent process tries to acquire a lock on account_move (e.g.  autovacuum),
it creates a deadlock situation where autovacuum is waiting for the lock
created by ALTER INDEX to be released, while ALTER TABLE gets in line waiting
for autovacuum to acquire and release it's lock. So we end the transaction
and get a new one, releasing the lock held because of ALTER INDEX before adding
the column.

closes odoo/upgrade#7338

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>