[FIX] base/0.0.0: handle collision on assigning temp code
In #7297, we assigned temp code to the countries which were missing
code. There might be already countries with code one
of `01, 02, 03, ... 99`. So if there are 2 custom countries with
code NULL and `01` we will get **psycopg2.errors.UniqueViolation**
when upgrade tries to set `01` to the NULL code.

For fixing the issue I started setting the temp from already existing
maximum integer code +1 instead of `01`.

Yes we are sure we will get issue if cliet have country with code=`99`
but in that case we want the upgrade to be blocked and the issue handled
manually.

tbg-1880

```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/18.0/odoo/service/server.py", line 1318, in preload_registries
    registry = Registry.new(dbname, update_module=update_module)
  File "<decorator-gen-13>", line 2, in new
  File "/home/odoo/src/odoo/18.0/odoo/tools/func.py", line 97, in locked
    return func(inst, *args, **kwargs)
  File "/home/odoo/src/odoo/18.0/odoo/modules/registry.py", line 127, in new
    odoo.modules.load_modules(registry, force_demo, status, update_module)
  File "/home/odoo/src/odoo/18.0/odoo/modules/loading.py", line 422, in load_modules
    loaded_modules, processed_modules = load_module_graph(
  File "/home/odoo/src/odoo/18.0/odoo/modules/loading.py", line 181, in load_module_graph
    migrations.migrate_module(package, 'pre')
  File "/home/odoo/src/odoo/18.0/odoo/modules/migration.py", line 254, in migrate_module
    mod.migrate(self.cr, installed_version)
  File "/tmp/tmpm5uv2gyl/migrations/base/0.0.0/pre-res_country_code.py", line 6, in migrate
    cr.execute(
  File "/home/odoo/src/odoo/18.0/odoo/sql_db.py", line 354, in execute
    res = self._obj.execute(query, params)
psycopg2.errors.UniqueViolation: duplicate key value violates unique constraint "res_country_code_uniq"
DETAIL:  Key (code)=(01) already exists.
```

closes odoo/upgrade#7340

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>