[FIX] stock/saas~17.3: don't check company in write always
Due to an odd behaviour in `_check_company` we check the company of all
fields as long as just one field is written whose `check_company`
property was set to `True`. This is not only wasteful but it is
undesirable during upgrades because it causes unnecessary errors. If
clients had bad data and we are only writing in some fields we don't
need to trigger a full check in all fields. Involved records are already
wrong and probably clients don't care or could fix it in some other
ways. Note that if the company of the record is changed this would still
trigger a check of all fields as expected.
https://github.com/odoo/odoo/blob/4cb362fcc3c2ba2489ef1de7c20c9a4fd8921ab8/odoo/models.py#L4336-L4337

This behaviour was already fixed in master in odoo/odoo#202060

The wrong data observed in failures is when some warehouses have an
address (`partner_id`) that belongs to a different company. Example:
```
yaih@(none):yaih_2400220> SELECT sw.id, sw.name, sw.company_id, sw.partner_id, p.company_id FROM stock_warehouse sw JOIN res_partner p ON sw.partner_id = p.id WHERE sw.id = 12
+----+--------------------------------------+------------+------------+------------+
| id | name                                 | company_id | partner_id | company_id |
|----+--------------------------------------+------------+------------+------------|
| 12 | NEUTRON SUPPLY, INC. - warehouse # 3 | 4          | 1          | 1          |
+----+--------------------------------------+------------+------------+------------+
```
```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/18.0/odoo/service/server.py", line 1306, in preload_registries
    registry = Registry.new(dbname, update_module=update_module)
  File "<decorator-gen-13>", line 2, in new
  File "/home/odoo/src/odoo/18.0/odoo/tools/func.py", line 97, in locked
    return func(inst, *args, **kwargs)
  File "/home/odoo/src/odoo/18.0/odoo/modules/registry.py", line 127, in new
    odoo.modules.load_modules(registry, force_demo, status, update_module)
  File "/home/odoo/src/odoo/18.0/odoo/modules/loading.py", line 514, in load_modules
    migrations.migrate_module(package, 'end')
  File "/home/odoo/src/odoo/18.0/odoo/modules/migration.py", line 254, in migrate_module
    mod.migrate(self.cr, installed_version)
  File "/tmp/tmpnol0ddg2/migrations/stock/saas~17.3.1.1/end-migrate.py", line 18, in migrate
    warehouse.write(new_types_vals)
  File "/home/odoo/src/odoo/18.0/addons/mrp_subcontracting/models/stock_warehouse.py", line 37, in write
    res = super().write(vals)
  File "/home/odoo/src/odoo/18.0/addons/mrp/models/stock_warehouse.py", line 263, in write
    return super(StockWarehouse, self).write(vals)
  File "/home/odoo/src/odoo/18.0/addons/stock/models/stock_warehouse.py", line 215, in write
    res = super().write(vals)
  File "/home/odoo/src/odoo/18.0/odoo/models.py", line 4797, in write
    self._check_company()
  File "/home/odoo/src/odoo/18.0/odoo/models.py", line 4363, in _check_company
    raise UserError("\n".join(lines))
odoo.exceptions.UserError: Incompatible companies on records:
- “NEUTRON SUPPLY, INC. - warehouse # 3” belongs to company “NEUTRON SUPPLY, INC.” and “Address” (partner_id: 'AKSHAR ENERGY SOLUTIONS, INC') belongs to another company.
```

In this file we write in other records, not `partner_id` (example from a
debugging session):
```
(Pdb) p vals
{'qc_type_id': 28, 'store_type_id': 29, 'xdock_type_id': 30, 'write_uid': 1, 'write_date': datetime.datetime(2025, 3, 14, 19, 15, 7, 407468)}
```

tbg-1903

closes odoo/upgrade#7361

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>
Co-authored-by: Alvaro Fuentes <afu@odoo.com>