[FIX] hr_payroll/saas~18.1.1.0: create salary rule category
In version saas-18.1, the country_id field is introduced for
hr_salary_rule_category, also constraints were added for the same, i.e,
related salary structure, the country and salary category's country should be
the same. This changes block the issue while set a country_id in the salary rule
category in the standard records due to a mismatch of the country between salary
structure and salary rule category.

To resolve the traceback, override the '_check_rule_category_country' method and
The list of mismatched records is given to the customer through the report.

opw-4563712, 4583302
upg-2561853, 2497676
tgb-1811

Part-of: odoo/upgrade#7375
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>