[FIX] hr_recruitment/saas~17.5: Keep attachments on applicants
In Odoo 18 the model `applicant` is split in models `applicant` and `candidate`,
where a candidate represents the "real person" and applicant is the link between
the candidate and a job position.

During the upgrade, the attachments are moved from the existing applicants to
the newly created candidates. This can create an issue, as it looks like the
attachments have dissappeared from the records.

Since while creating a new applicant, the attachments from the corresponding
candidate are copied to it, the existing attachments should also be copied.

opw-4592268

Part-of: odoo/upgrade#7374
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>