[FIX] l10n_nl_reports_sbr/saas~17.5: Fix NL keys and certificates
Problem
---------
1. During migration, the keys' passwords are not migrated properly.
This is because the password is 'stored' in 'res.config.settings' -
transient table. For that reason, the password is actually stored in
'res.company'.
2. The PKI and root certificate IDs are meant to be automatically set up
on the company record during the upgrade using a JOIN table to simplify
the request. However, in some cases, one out the two certificate is
present making the (INNER) JOIN on the company_id empty in such case.

Solution
---------
1. Select the password value from the 'res.company' table rather than
'res.config.settings' and delete the column afterwards.
2. Use an FULL JOIN so that if any of the PKI or root certificate is
being migrated, its ID will be linked.

upg-4643130

closes odoo/upgrade#7380

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>