[FIX] documents/18.0,saas~18.{1,2,3}: fix thumbnail status
In v17, _compute_thumbnail_status write a False status when the thumbnail is
updated because when it is triggered the attachment is not yet created. So the
thumbnail is present but its status is incorrect. It works in v17 because the
status is ignored by the view for images but in v17.4 and further version, the
status is not ignored and the thumbnail is then not displayed.

In this upgrade we set thumbnail_status to "present" for all document thumbnail
present in ir_attachment with a thumbnail_status value of False. We do it for
the upgrade from 17.4 to 18.3 as the problem has been detected when developing
18.3 and database already upgraded to those versions have potentially wrong
status written by the v17.

For database already upgraded, we solve the problem by a temporary fix in
odoo/enterprise#81558.

Task-4642866

closes odoo/upgrade#7392

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>