[FIX] account_budget/saas~17.3: properly migrate old account_id data
this commit c188f65, creates the analytical plan's column field in the model
`budget.line` and set value according to the plan for old data of `account_id`
field.

issue:
---------
Data is not migrating properly for sub plans.

cause of the issue:
-----------------------
We are fetching root analytic plans to create columns. However, during migration
, we are using the same values as `plan_id` instead of `root_plan_id` values
in the `account.analytic.account` model.

steps to reproduce:
------------------------
1. Create a sub plan and a related analytic account in 17.0.
2. Create budget lines using the sub plan.
3. Migrate it to 18.0.

upg-2592814
opw-4614068

closes odoo/upgrade#7400

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>