[FIX] sale_subscription/0.0.0: remove field for further versions
The fields on sale.subscription.report inherited from
sale.report are marked as coming from the
sale_subscription module, so when the fields are
removed this will not be reflected on the child model.
There are currently failures in higher versions because
the website_sale module is uninstalled and the field
on the subscription report is not removed.

closes odoo/upgrade#7401

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>