[FIX] l10n_in_edi/saas~18.2: correct the xml_id
In this commit
We fix the typo in the xml_id of `account.edi.format` which
must be removed after the migration. The issue was not identified
because unless someone less tried to make the record `noupdate=1`
Due to which record was not removed.

A fix of the PR-https://github.com/odoo/upgrade/pull/7098

see https://github.com/odoo/odoo/pull/191920
see https://github.com/odoo/enterprise/pull/76621

closes odoo/upgrade#7406

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>