[IMP] base/saas~17.5: remove consolidation, and report user options
The consolidation app is removed in odoo 18 and the upgrade is
blocked unless the env variable to remove it set.
Now the variable will be set by default and the upgrade will be
blocked only if it is explicitly set to False.
The upgrade report offers more information and guidlines to users
who used the consolidation app on what to do to recover and how to
adapt to the new way, with a link to the odoo documentation.

opw-4655592

closes odoo/upgrade#7415

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>