[ADD] base_automation/saas~18.3: after/before mode for time trigger
This upgrade script make adaptations to the added delay mode for
time-based automations.

Before [1], it was possible to create a time-based automation with a
negative delay value, which would trigger the automation before the
specified time. After [1], the delay value is always positive and a new
selection field is added to allow to trigger the automation after/before
the specified time. This script updates the existing automations to
reflect the new behavior.

Task id: opw-4648932

[1]: https://github.com/odoo/odoo/pull/202347

closes odoo/upgrade#7431

Related: odoo/odoo#202347
Related: odoo/enterprise#82050
Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>