[IMP] spreadsheet_edition/saas~18.4: add 'cumulatedStart' to charts
Since 17.0, we wrongly assumed that the properties 'cumulative' and
'cumulatedStart' of a chart view were equivalent. This proved to be
false and we now correctly make the distinction. Since users might have
designed their spreadsheet around this issue, we add a fallba

This upgrade PR populates the 'cumulatedStart' key of the chart-related
commands to mirror the old behaviour and future spreadsheets and
their commands will contain the correct distinction.

In stable versions, since ysers might have designed their spreadsheet
around that issue, we add a fallback for existing commands but the new
created commands will have the distinction and this PR takes that
behaviour into account as well.

https://github.com/odoo/odoo/pull/205117 - adds cumulatedStart
distinction to chart definition
https://github.com/odoo/enterprise/pull/83018 - adds edition of the
property

task-4701303

closes odoo/upgrade#7506

Related: odoo/odoo#205117
Related: odoo/enterprise#83018
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>