[FIX] link_tracker/saas~14.4: remove sql constraint
commit b7e5eb639abb5b3a170f97f58f2cad3099b309df removed the unique
constraint on link_tracker, but there were no upgrade scripts related.

This leads to an error when a user upgrades from a previous version, then
tries to have the same link in a marketing campaign on multiple objects
(e.g. an image that links to the campaign page via the same link as the
call to action button).

opw-4680129

See odoo/odoo#48845
See odoo/enterprise#9608

closes odoo/upgrade#7529

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>