[IMP] documents_project/saas~18.1: Remove 100 documents limit
The documents linked to projects or tasks but not held in the corresponding
project folder or its subfolders create shortcuts in the project folder.
Since it was unknown how many documents would be affected by this, a cap of 100
was set, failing the upgrade.

Most of the cases seem related to reorganization of the documents.

This commit removes that hard limit. To avoid creating too many shortcuts on the
same folder, a new subfolder might be created to hold the shortcuts.

tbg-1785

closes odoo/upgrade#7514

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>