[IMP] documents_hr{_payroll}/saas~18.3: switch Payroll for Employees
With odoo/enterprise#80399 and odoo/odoo#203729,
the 'Payroll' folder has been deprecated while an 'Employees'
folder is introduced, in which a subfolder is created for each
employee.
Those folders are not meant to store payroll documents, as
payroll documents will now be stored on each related employee's
Drive. Employees folder and subfolders are meant to store HR
documents related to employee (that are not payroll).

All the documents that were inside Payroll folder are redistributed
to their related employee's 'My Drive' to fit the new paradigm in
which payroll officers do not have access to the payroll documents
anymore. They still have access to the attachment on which documents
are based.

Task-4309991

closes odoo/upgrade#7532

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: Florian Charlier <flch@odoo.com>