[FIX] base/saas~16.1: create team_id column
The module 'sale' is a new dependency for 'spreadsheet_dashboard_hr_expense'
and it could be force installed during the upgrade.
The field 'account.move.team_id' in the module 'sale' is a stored
computed field, and it is also tracked. The account.move modified records
tend to be numerous, and while the computation itself may not cause a
memoryerror, the tracking logs will be problematic since they are all
created and written on a flush for all the affected records.
Creating the column before loading the module will skip the computation
and the tracking.

closes odoo/upgrade#7541

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>