[FIX] account_reports/0.0.0: Replace dash with tilde
- **Description**

From version 16.0 onwards, the (~) tilde format became the standard for line_id.
Additionally, starting from SaaS~17.4, the account_report_footnote model was
renamed to account_report_annotation, and several improvements were made to the
annotation system.

- **Before fix**

In version 17.0, footnote rendering was done using the get_footnotes method,
which allowed footnotes even if the line_id started with a dash (-) and didn’t
strictly check the format. However, annotations with an incorrect line_id format
with a (-) dash were not displayed in the corresponding report. Although
the report loaded without errors, and those annotations were silently skipped
and didn’t appear.

However, in saas~17.4, a new method get_annotations was introduced, which
includes stricter validation logic. It expects the line_id to use the ~ (tilde)
format. If this format is not followed, an error is raised when accessing the
report.

- **After fix**

This fix ensures compatibility with annotation formats from pre-17.4 versions by
converting all line_id values that use the dash (-) to the expected tilde (~)
format.This helps prevent runtime errors and ensures annotations render
correctly.

Steps to Reproduce:

- Create a database in Odoo 15.0 and add footnotes to a Balance Sheet report.
- Migrate the database to saas~17.4.
- Open the report where footnotes were previously added.
- The annotations (footnotes) will no longer render because of the stricter
  line_id validation, and a traceback is also generated.

closes odoo/upgrade#7542

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>