[IMP] sale_amazon/saas~18.4: transfer is_fbm to amazon_channel
Following odoo/enterprise#78897, the `amazon_feed_ref` field has been
repurposed to store additional information needed for the inventory
feed. This commit is the start of a chain to clean up this change and
store each new info in its own field.

odoo/enterprise#78293 added a new `amazon_channel` field which is
similar to the `is_fbm` key previously stored in `amazon_feed_ref`. With
this commit, the key is removed and transferred to the new field.

task-3820591

closes odoo/upgrade#7507

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>