[IMP] sale/saas~17.5: set analytic distribution on sol
The "analytical_account_id" field has been removed from sale_order in version
saas~17.5. Sales orders created in the older version with analytic accounts (AA)
do not carry over the AA to invoice lines after upgrading to version 18.0.
As a result, the gross margin associated with these AAs does not get updated
when new invoices are created. To tackle that, set the AA on the analytic
distribution field of SOLs (if the analytic distribution is empty) during the
migration to be sure the analytic distribution field on account move is set as
expected.

Unicorn task - 4731167
upg - 2459870, 2648064

closes odoo/upgrade#7582

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>