[FIX] onboarding/18.0: Remove progress with wrong company value
The onboarding progress is supposed to be per company or global. If at least one
of the onboarding steps is made by company, the whole onboarding is.

It looks like through a bug some dbs managed to have progress with and without
a company set for the same onboarding. This triggers an error when two records
are fetched and only one expected.

There was a patch in standard to accomodate for this case but it was only for
saas~16.4. https://github.com/odoo/odoo/pull/108183

tbg-1896

```
ValueError: Expected singleton: onboarding.progress(1, 2)
````

closes odoo/upgrade#7583

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>