[IMP] event{_sale}/saas~18.3: update to multi slots
Following the addition of the multi slots events, the mailing
templates need to be updated so that the slot datetimes can be
displayed instead of the event ones for multi slots events.

Renaming the 'event_ticket_id_change_exception' so that it can be
used for both a ticket change or a slot change on the sale order.

Create the column for the new stored computed field
"sale.order.line.event_slot_id" to avoid their computation by
the ORM.

related:
odoo/odoo#205945
odoo/enterprise#83411

Task-4307566

closes odoo/upgrade#7565

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>