[FIX] account/saas~18.{1,2,3}: recreate index
Originally the index was named `account_move_line__unreconciled_index`.
After the refactor in odoo/odoo#174829 the name is now generated from an
`Index` object. This keeps a single underscore between the table and the
index name. Thus the new index name is
`account_move_line_unreconciled_index`.

Renaming the index is not enough since the ORM relies on the associated
COMMENT to decide later if the index needs to be updated. This finally
happens in 18.3. Thus we just remove the index and let the ORM re-create
it. Moreover DBs in 18.1+ that underwent an upgrade have both indexes.

closes odoo/upgrade#7596

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>