[FIX] documents_hr_recruitment/18.0: Ensure parent folder is correct
The record `document_recruitment_folder` is changed in 18.0 to be moved to the
folder `document_hr_folder`. However that folder can be archived or the change
can create a loop, so we add a check before the update.

upg-2709912
upg-2646254
```
raise UserError(_("It is not possible to move archived documents or documents to archived folders."))

```
upg-2627419
```
raise UserError(_("Recursion Detected."))
```

closes odoo/upgrade#7613

Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>