[FIX] l10n_nl/18.0: ignore force_create flag
The `force_create` flag in `try_loading` is problematic because it skips
the creating/update of many elements when loading a CoA. Since
odoo/odoo@a3c57b4f we partially load data in l10n_nl. This leads to
errors during the upgrade. The patch in stable was meant to skip
overriding possible customization by clients in the data loaded by the
CoA. During a major upgrade we can force the override because clients
are supposed to test their upgrade and verify that all is properly
configured.

tbg-1952

Part-of: odoo/upgrade#7615
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>