[IMP] pos_settle_due/saas~18.3: Populate new fields and avoid computes
Some new fields are introduced. To avoid computes by the ORM, they are created
beforehand by the script.

The field `settle_invoice_product_id` is populated by a function at install,
but it's not run during an upgrade, so to keep it consistent we update the value
with a query.

closes odoo/upgrade#7630

See: https://github.com/odoo/enterprise/pull/80267
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>