[FIX] hr_recruitment/0.0.0: move actions to `hr_recruitment_sms`
Odoo: https://github.com/odoo/odoo/pull/199238

The bridge module between `sms` and `hr_recruitment` has been
backported starting from version `17.0`. So we need to move these
actions to bridge module `hr_recruitment_sms` in **0.0.0** script:
1. `hr_recruitment.action_hr_applicant_mass_sms`[^1] was introduced
   in `saas~16.4`
2. `hr_recruitment.action_hr_candidate_mass_sms`[^2] was introduced
   in `18.0` and being removed in `saas~18.2`[^3]

So I renamed the xmlid of first action if the version is between
`[17.0, 19.0]` because we need to ensure the script run for all
upgrades from >= `16.4` and targeting <= `19.0`. Additionlly
I renamed the xmlid of second action if the version is
equal to `18.0` or `saas~18.1`.

[^1]: odoo/odoo@5da9a87
[^2]: https://github.com/odoo/odoo/blob/e5b2a1744ff2d2ab4fec8b79256519107e561bcd/addons/hr_recruitment/views/hr_candidate_views.xml#L220-L232
[^3]: odoo/odoo#189274

closes odoo/upgrade#7646

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>