[FIX] documents/saas~18.{1,3}
In `saas~18.1`, a new folder has been added. As it is declared with
`forcecreate=0`, we need to create it using an upgrade script (the
`if_unchange` was useless as the record is not created by the normal
loading).
We however need to reset the `noupdate` flag of the newly created
records to match the XML declaration.

It wasn't a problem during the `saas~18.2` upgrade, but in `saas~18.3`,
the alias on documents is now optional. Technically, this mean that
`documents.document` doesn't "_inherits" from `mail.alias` anymore
(`alias_id` is now a plain m2o in ON DELETE RESTRICT).
This has the side effect of not flagging the xmlid of the alias as seen
when loading the XML file and thus trying to delete it at the end of the
upgrade (which fail).

Part-of: odoo/upgrade#7694
Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>