[IMP] product/11.0: Precompute new columns to speed up upgrade
Computed fields are pre-computed to avoid the ORM doing it, so instead of hours
it can take virtually nothing if no mail.activity records exist.

The affected database has close to 10M product.product records
but 0 mail.activity and it still takes 8 hours for the ORM to do the computing.

upg-2695637
opw-4774880

closes odoo/upgrade#7649

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>