[FIX] base/0.0.0: set source and target env vars
The env variables 'ODOO_UPG_DB_SOURCE_VERSION' and
'ODOO_UPG_DB_TARGET_VERSION' which are used few times
in the upgrade are set in 2 places: the upgrade platform
and in upgradeci/matt. The runbot tests run without
setting these variables, and we may get undesirable
effects.
If we are on CI we can assume we are not doing multi
step upgrades, and the current version is the source
while the release version is the target.

closes odoo/upgrade#7660

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>