[FIX] base/0.0.0: ensure superuser has active company
In the customer database, two companies existed—
one inactive and one active. The superuser was linked
only to the inactive one, which led to failures in
reporting logic during the upgrade from 17.0 to 18.2,
due to missing accessible companies in context.

To prevent such issues in future versions (e.g. 18.3),
the condition for assigning companies to the root user
was updated to include all target versions >= 17.0,
ensuring the superuser always has atleast one active
company access.

closes odoo/upgrade#7669

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>