[FIX] hr_expense/saas~18.2: remove duplicate attachments
Prior to 18.2, when adding an attachment directly on an hr.expense
record (that is linked to an hr.expense.sheet), it would also add
the same attachment to its corresponding hr.expense.sheet.

Since in 18.2 we are merging the two models, we need to remove those
duplicates from hr.expense.sheet, or else we will end up with
duplicated attachments on the hr.expense model after the merge.

closes odoo/upgrade#7692

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>