[IMP] l10n_ae/saas~18.4: Update UAE accounting configurations
The tax report does not use the new reporting engine from v16 (it is
currently too long and unclear to read); Sales export and Purchase
Import taxes are set up incorrectly in respect of their tax grids, and
the naming of the tax groups is unclear. We take this opportunity to
overhaul the module and give it a new look in preparation for its
e-invoicing phase in 2026.

- Update the tax report in account_tax_report_data.xml to use the new
  tax grids, labels and three columns (for base, tax and adjustment)
  instead of 2 sections (base and tax) which made the report very
  long.
- Modified account names and types in account.account-ae.csv
- Adjusted tax definitions in account.tax-ae.csv to reflect updated
  naming conventions and added new taxes for postponed payments and
  reverse charge provisions, also update tax grids (names but also
  distribution).
- Adding reverse charge and out of scope tax groups in
  account.tax.group-ae.csv.
- Set Dubai tax as a default instead of Abu Dhabi in template_ae.py.

task-4741917
Community PR: https://github.com/odoo/odoo/pull/210853

closes odoo/upgrade#7714

Related: odoo/odoo#210853
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>