[IMP] website_crm_partner_assign/saas~18.3: fix upgrade wo/ membership
On a database with `website_crm_partner_assign` module installed but
where `membership` is not installed, the upgrade script of `partnership`
module were not run because this module was considered `to install`.

This commit force running upgrade script in that case, and split the
part related to the `membership` and `website_crm_partner_assign` to
make it clearer.

closes odoo/upgrade#7716

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>