[FIX] project_mrp/17.5.1.0: create/write metadata for analytic projects
The analytic projects created to replace removed analytic account fields
didn't have the field create_date, create_uid, write_date and write_uid
set. This caused a traceback when opening the project if project stages
were enabled in the settings.

This fix sets the current time for the record creation and OdooBot as
the user.

opw-4809303
upg-2804568

closes odoo/upgrade#7780

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>