[IMP] project_mrp/17.5.1.0: human-readable names for analytic projects
Analytic project names are currently of the form
"Analytic Project (137,)", with the number being the record id of the
analytic account the project is linked with. Since the analytic account
field got replaced with the project field on records like sales and
manufacturing orders, this can be confusing to users because they can't
see anymore at a glance which analytic account the field is referring
to.

This change uses the name of the analytic account instead of its id. If
there are multiple they are separated by a '|' like analytic
distributions with multiple accounts.

opw-4809303
upg-2804568

closes odoo/upgrade#7782

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>