[FIX] account/0.0.0: update account types
A new validation on accounts used in tax groups is added.
It runs when a tax group is updated through try_loading during
the upgrade, however the existing accounts are not updated.
Here we intercept the write before the validation and update
the needed accounts. Only standard accounts are updated.

https://github.com/odoo/odoo/pull/201249

closes odoo/upgrade#7787

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>