[IMP] spreadsheet_edition/saas~18.4: ensure that fixedPeriod is correct
Some migrations (for example `antepenultimate_year`) introduced a
fixedPeriod with a yearOffset different from 0, which was not
expected by the global filters side panel. (It worked fine by chance in
the model).

This commit ensures that the fixedPeriod is always correct when the
migration is applied.

See https://github.com/odoo/odoo/pull/212884

Task-4848233

closes odoo/upgrade#7789

Related: odoo/odoo#212884
Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>