[FIX] documents/saas~17.5: fix account multicompany and modified rules
Account workflow rules used company properties to store a set
journal_id, but were otherwise able to infer a journal_id based
on the context when executed.

We should therefore, in upgrade scripts, create (and embed)
server actions as:

* one for each company using a fixed journal_id,
* one that should work with any company

The last one will use the dynamically computed
`journal_id` so that it is available to companies that
never called this method yet (including the eventual new
companies that won't have to duplicate the action).

Also, we cannot assume that modified records serve the
same purpose as the original records, so we should
migrate them as well.

Task-4827396

closes odoo/upgrade#7804

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>
Co-authored-by: Pierre-Yves Dufays <pydu@odoo.com>