[FIX] account_disallowed_expenses_fleet/saas~18.2.1.0: deductibility
Our Disallowed Expense module does not allow for deductibility of >100%.
We have cases where it's needed, for example in Belgium you are allowed to
deduct 120% of your amounts expensed for eInvoicing. For this we need to inverse
the percentage to the part that is allowed in the expenses, so that we can take
it into account in the report, and add another column showing the amounts that
can actually be deduced.

This was already done for the `account_disallowed_expenses` module but not
for `account_disallowed_expenses_fleet`. This commit fixes this.

Entreprise PR: odoo/enterprise#76066
Documentation PR: odoo/documentation#11634

task-id 4421297

closes odoo/upgrade#7807

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>