[FIX] base/saas~18.4: add HTML properties
Purpose
=======
We added HTML properties. This type of field is sanitized when writing,
and so, we expect the values in database to be safe.

The HTML properties have a convention which is that their names need
to end with `_html` (it allows us to make some optimization).

But, existing non-html properties could already exist with a name
ending with `_html` (if crafted by a malicious user), and so we
need to clean the value in database (if no malicious user created
property, we expect to update no record).

For simplicity, we don't try to rename the properties, we just set
the value to false (since it's not possible for a non-malicious user
to create one).

Task-4712395
See odoo/odoo/pull/212175
See odoo/enterprise/pull/86751

closes odoo/upgrade#7818

Related: odoo/odoo#212175
Related: odoo/enterprise#86751
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>