[IMP] im_livechat/saas~18.4: block chat assignment when in call
The community counterpart of this commit creates a new live chat option
used to block assignment of chats for agents who are in a call. Before,
this behavior was coupled to the `max_session_mode` field being set to
`limited`. This commit fills the new `block_assignment_during_call`
column according to the value of `max_sessions_mode`.

task-4818793

closes odoo/upgrade#7814

Communtiy: odoo/odoo#213148
Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>