[FIX] account_budget/saas~17.3: fix analytic distribution
Since the following commit: 193f10b
the use of JSONB_EACH_TEXT in the migration script caused the
analytic_distribution values to be incorrectly stored as strings
instead of floats.

This fix replaces JSONB_EACH_TEXT with JSONB_EACH to preserve the
correct float values.

Steps to reproduce:
- Start from version 17.2
- Create a budget position and assign it to account “TEST”
- Create a journal entry with that account and an analytic distribution
- Upgrade to version 17.3
- The stored analytic_distribution values will be strings instead of floats

This will cause subsequent issues, since we will try to round the string
value, instead of the float value, leading to TypeError

opw-4816999
opw-4840143
opw-4826929

closes odoo/upgrade#7829

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>