[IMP] stock/saas~18.5: replace picking `date` with `create_date`
This change was suggested by a contributor, but it requires an upgrade
script, so a set of 3 new PRs was created. The original purpose was to
improve the stock picking filtering in `simulate_period` test helper
method. While working on this, it became apparent that the `date` field
in stock picking is useless and it should be replaced by `create_date`.

Related:
odoo/odoo#213949
odoo/enterprise#87494
odoo/upgrade#7852

Original community PRs:
odoo/odoo#171895
odoo/enterprise#66084

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>