[IMP] account_reports/saas~18.4: remove summary and make message char
The PR in enterprise, adds some improvements for `account.return.check`.
Like adding a form view, tracking on some fields and add activities.
In the process, there are some changes that requires an upgrade:

The field `summary` of `account.return` is removed.

The field `message` was Html but as it serves no purpose, it is now a
Char.

Enterprise PR: odoo/enterprise#86707

task-4823849

closes odoo/upgrade#7848

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>