[IMP] hr_recruitment/saas~18.5 update degrees
The related community/enterprise PR's introduces new and improved
behaviours for managing skills in relation to recruitment.
A part of that improvement was adding a score field to the
hr.recruitment.degree model to allow for job matching calculations.
As a result of that, the existing degrees records need to be updated to
get the new score value.
Another part of that improvement was adding fully fledge skill requirements
to a job position, so a skill and level rather than just a skill in the form
of hr.job.skill. A side effect of that change is that the skill_ids field on
jobs is now computed and the meaning of it is slightly different so this pr
creates a hr.job.skill for each skill in the skill_ids field.
Additionally, we remove the unused field currency_id and compensation on
hr.job and a form view form hr_recruitment_skills.

Community PR: https://github.com/odoo/odoo/pull/205542
Enterprise PR: https://github.com/odoo/enterprise/pull/83225

closes odoo/upgrade#7858

Related: odoo/odoo#205542
Related: odoo/enterprise#83225
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>