[FIX] product/0.0.0: do not update write_date
This is to avoid `MemoryError`s

In 17 we have a compute to update write_date in products whenever the
write_date of their templates was updated.

https://github.com/odoo/odoo/blob/0ad569a7e1cfddcf238f5cec7a9e303d89df2a67/addons/product/models/product_product.py#L133

This could lead to memory overrun if during the upgrade a bunch of
templates get updated that subsequently trigger many more re-computes in
products. Here we skip the whole compute and do it in SQL in the end.

See odoo/odoo#204931

closes odoo/upgrade#7866

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>